requirements.txt
python-binance>=1.0.19
python-dotenv>=1.0.0
typer[all]>=0.9.0
rich>=13.7.0
questionary>=2.0.1
